requirements.txt
# CAO Data Extraction Pipeline - Dependencies
# Generated automatically by scanning all Python imports in the repository

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# PDF processing and OCR
PyPDF2>=3.0.0
pdf2image>=3.1.0
pytesseract>=0.3.10
pdfplumber>=0.10.0

# Web scraping and HTTP
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0

# Google AI / Gemini API
google-generativeai>=0.3.0
google-genai>=0.1.0

# Data validation and type checking
pydantic>=2.5.0

# Environment management
python-dotenv>=1.0.0

# Translation services
deep-translator>=1.11.0

# Data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Document processing
python-docx>=1.1.0

# File locking (for parallel processing)
# fcntl is built-in on Unix systems, no package needed

# Development and testing utilities
astor>=0.8.1
